Remove 'args' from modules
Using 'args' as an argument has the advantage of easily passing
multiple arguments/options in one argument, but 'args' is a bit too
much tied to argparse. In the future, a dict may be used to take over
this functionality, since long argument lists are not always nice to
work with.