Initial re-organisation of the code
This re-organises the interal code for GOTOtile in a slightly more
object oriented manner. This should make it clearer (easier) how to
use GOTOtile as a module in other code.

The underlying functionality is exactly the same, and currently nearly
all calculations are still done with the original functions.

The new API is implemented alongside the original code, and both can
be used with the same code base. As for the scripts: tileskymap uses
the old interface, while tileskymap-new uses the newer interface. The
latter is still incomplete, since the plotting output section still
needs work.

The new interface also offers one extra option: it can use multiple
cores to calculate which tiles are visible (i.e., what is used with
the --nightsky option). This significantly reduces the computation
time for that part of the code (which in general is the bottleneck in
a tileskymap run).