Generalize the galaxy catalog option
The option to fold with the GWGC catalog is changed to let a user
specify any catalog. The option is therefore renamed to --catalog, and
comes with an extra option --catalog-weight-key (see below). The
'galtools' module has been renamed 'catalog'.

The catalog file given by a user has a few constraints:
- it should be readable by astropy.table.Table.read(...)
- it should contain an 'ra' and a 'dec' column, both floating point
  values of degrees

By default, no extra weighting is done, and the skymap is folded with
the galaxy density. When --catalog-weight-key is set to a column name
in the user-supplied table, this column is used as the weight.

Note that when the --catalog option is given just by itself, with no
user-supplied file, the built-in GWGC catalog is used, and the skymap
is weighted by the 'Blum' column.