Allow for a user telescope configuration file
This change allows a user to specify one or more telescopes through a
configuration file in YAML syntax.

This removes the --site and --fov options. If a user wants to override
an existing telescope with some changes, this has to be done
explicitly in the configuration file, and all the site and
field-of-view parameters have to be given.

An example syntax is as follow:

- short: 2p2wfi
  full: ESO/MPI 2.2m WFI
  longitude: -70.7313
  latitude: -29.2612
  height: 2400
  fov-ra: 0.567  # degree
  fov-dec: 0.55   # degree
- short: vista
  full: VISTA
  longitude: -70.404
  latitude: -24.627
  height: 2518
  fov-ra: 1.5
  fov-dec: 1