Don't count insignificant tiles
Prevent the counting of tiles with very low probability
(insignificant; MIPROB = 1e-6) towards the maximum number of tiles
requested (--maxtiles option).

Previously, each iteration where it was night for at least one
telescope, the number of observed tiles was increased. In rare cases,
however, the only visible patches on the sky may have been of very low
significance, and should be ignored. The end result would be an empty
list of pointings.

Now, these "empty" observations are skipped, and only significant
tiles are counted. This may result in a list of pointings starting
later than the requested starting date.