Handle IERS-A server unavailability
When the IERS-A server is not available and astropy wants time
corrections (for e.g. coordinate transformations), it will attempt to
download the data for the duration of the timeout, then give up and
resort to the IERS-B file locally.

This is fine for a one-off, but is a pain in the arse if done for a
multitudes of observer frames. Unfortunately,
``astropy.utils.iers.conf.auto_download = False`` or
``astropy.utils.iers.conf.remote_timeout = 0.1`` seem to be ignored.

Instead, the work-aroud appears to set
``astropy.utils.iers.conf.iers_auto_url`` to something *not* a URL (''
or None): this will immediately cause an exception in the download
part of the IERS module, which is caught (any exception) and ignored
(just a warning issued), and the IERS-B file is read.

The ``utils.test_iers`` function tests for the IERS-A URL, and if that
(or a backup server) fail, it does just that: it sets the URL to a
non-URL value.

In addition, we set ``astropy.utils.iers.conf.iers_auto_url`` for each
indendent process when multiprocessing in
``skymaptools.get_visiblemap``. I haven't tested this properly, but if
a new Python instance is created, the
``astropy.utils.iers.conf.iers_auto_url`` value may also be reset for
each separate process.