Adjust sidereal times when timespan extends into the next night
Sideral times were calculated only for the current night. When the
given timespan extends into the next night, that part was completley
ignored.

The code will now calculate sidereal timespans for each interval when
it's night, up to the time allowed by `timespan` (and up to a maximum
of a year: a timespan of 100 years is regarded as identical to 1
year).