Apply the settings constants more flexibly
Using `getattr` grabs the (possibly changed) value of a constant in
the `settings` module at that moment, instead of at the time of the
import (which usually happens at module level). This allows an
external program to change constants in `gototile.settings` and thus
alter gototile's behaviour, without having to pass these constants as
arguments.

Note: the parser module does not use this, as it is not intended to be
called by an external program.